Move loadCJS out of core build to desktop
