+++ products/jbrowse-desktop/src/StartScreen/util.tsx
@@ -3,6 +3,11 @@ import PluginManager from '@jbrowse/core/PluginManager'
 import PluginLoader from '@jbrowse/core/PluginLoader'
 import { readConfObject } from '@jbrowse/core/configuration'
 import deepmerge from 'deepmerge'
+import sanitize from 'sanitize-filename'
+
+export interface LoadedPlugin {
+  default: PluginConstructor
+}
 
 import {
   writeAWSAnalytics,
@@ -57,6 +62,48 @@ export async function createPluginManager(
 ) {
   const pluginLoader = new PluginLoader(configSnapshot.plugins, {
     fetchESM: url => import(/* webpackIgnore:true */ url),
+    fetchCJS: async url => {
+      const fs: typeof import('fs') = require('fs')
+      const path: typeof import('path') = require('path')
+      const os: typeof import('os') = require('os')
+      const http: typeof import('http') = require('http')
+      const fsPromises = fs.promises
+      // On macOS `os.tmpdir()` returns the path to a symlink, see:
+      // https://github.com/nodejs/node/issues/11422
+      const tmpDir = await fsPromises.mkdtemp(
+        path.join(await fsPromises.realpath(os.tmpdir()), 'jbrowse-plugin-'),
+      )
+      let plugin: LoadedPlugin | undefined = undefined
+      try {
+        const pluginLocation = path.join(tmpDir, sanitize(url))
+        const pluginLocationRelative = path.relative('.', pluginLocation)
+
+        await new Promise<void>((resolve, reject) => {
+          const file = fs.createWriteStream(pluginLocation)
+          http
+            .get(url, response => {
+              response.pipe(file)
+              file.on('finish', () => {
+                resolve()
+              })
+            })
+            .on('error', err => {
+              fs.unlinkSync(pluginLocation)
+              reject(err)
+            })
+        })
+        plugin = __non_webpack_require__(pluginLocationRelative) as
+          | LoadedPlugin
+          | undefined
+      } finally {
+        fsPromises.rmdir(tmpDir, { recursive: true })
+      }
+
+      if (!plugin) {
+        throw new Error(`Could not load CJS plugin: ${url}`)
+      }
+      return plugin
+    },
   })
   pluginLoader.installGlobalReExports(window)
   const runtimePlugins = await pluginLoader.load()
